feat: 해당 아이의 언어 수준 + 동일 나이 대의 아이 수준 비교 차트 구현

+++ src/main/java/soongsil/kidbean/server/summary/domain/AverageScore.java
@@ -36,17 +36,17 @@ public class AverageScore {
     private Long totalScore;
 
     @Column(name = "quiz_count")
-    private Integer memberCount;
+    private Long memberCount;
 
     @Builder
-    public AverageScore(AgeGroup ageGroup, Long sum, int memberCount, QuizCategory category) {
+    public AverageScore(AgeGroup ageGroup, Long sum, long memberCount, QuizCategory category) {
         this.ageGroup = ageGroup;
         this.totalScore = sum;
         this.memberCount = memberCount;
         this.quizCategory = category;
     }
 
-    public void updateScoreAndCount(long sum, int size) {
+    public void updateScoreAndCount(long sum, long size) {
         totalScore = sum;
         memberCount = size;
     }
